requirements.txt
google-api-python-client==2.103.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
pandas==2.1.1
openpyxl==3.1.2
pyyaml==6.0.1
python-dotenv==1.0.0
